Moved and renamed 3 module files to utilities.
Moved 3 module files to utilities and renamed them. Updated the rest of the files where these modules are called. These files are:
-PPGetLogger (now called sorchaGetLogger)
-PPCommandLinerParser (now called sorchaCommandLinerParser)
-PPModuleRNG (now called sorchaModuleRNG)
