+++ src/sorcha/modules/__init__.py
@@ -14,8 +14,6 @@ from . import PPOutput
 from . import PPRandomizeMeasurements
 from . import PPTrailingLoss
 from . import PPVignetting
-from . import PPCommandLineParser
-from . import PPGetLogger
 from . import PPApplyFOVFilter
 from . import PPSNRLimit
 from . import PPFadingFunctionFilter
