+++ src/sorcha/utilities/sorchaGetLogger.py
@@ -3,7 +3,7 @@ import os
 from datetime import datetime
 
 
-def PPGetLogger(
+def GetLogger(
     log_location,
     log_stem,
     log_format="%(asctime)s %(name)-12s %(levelname)-8s %(message)s ",
