+++ tests/sorcha/test_PPAddUncertainty.py
@@ -23,7 +23,7 @@ import numpy as np
 import pandas as pd
 from numpy.testing import assert_almost_equal, assert_equal
 from sorcha.utilities.sorchaConfigs import sorchaConfigs, expertConfigs
-from sorcha.modules.PPModuleRNG import PerModuleRNG
+from sorcha.utilities.sorchaModuleRNG import PerModuleRNG
 
 
 def test_calcAstrometricUncertainty():
